src/requirements.txt
flask
tensorflow-cpu
pillow
numpy
cv2